+++ CHANGES.md
@@ -1,6 +1,7 @@
 # 2.10.0 (Unreleased)
 - [NEW] Add IAM cookie authentication method.
 - [IMPROVED] Clarified documentation for search indexes.
+- [FIXED] Connection leaks in some session renewal error scenarios.
 
 # 2.9.0 (2017-04-26)
 - [NEW] Add faceted search variable argument to `drillDown` method allowing multiple drill down